backend/requirements.txt
fastapi
uvicorn
crewai>=1.0.0
langchain-openai
langchain-community
httpx
redis
pydantic>=2.0.0
pydantic-settings
python-dotenv
backoff
openai
loguru
slowapi
python-multipart
